requirements.txt
Django==1.6
South==0.8.4
celery==3.1.5
django-bower==4.8.1
django-celery==3.1.1
django-crispy-forms==1.4.0
django-pipeline==1.3.17
pysqlite==2.6.3
